Beacons Section Edits
Capitalization and links for the Beacons section.

TBD: In 6.2 Actors - we use the phrase aggregate BTC1 Beacon with only BTC1 Beacon linked. Should we add Aggregate BTC1 Beacon as a link option to Aggregate Beacon in terminology, or should they link separately to Aggregate and to BTC1 Beacon? 

TBD: In 6.6.2, we use the phrase SMT Beacon signal - how should we link for this one? Similar to the question above, and since there was no linking I added the SMT Beacon link for now, but that may not be sufficient.